requirements.txt
aiohappyeyeballs>=2.3.4
aiohttp>=3.10.1
aiosignal>=1.3.1
annotated-types>=0.7.0
anyio>=4.4.0
asgiref>=3.8.1
asttokens>=2.4.1
attrs>=24.1.0
autobahn>=24.4.2
Automat>=24.8.1
backoff>=2.2.1
bcrypt>=4.2.0
beautifulsoup4>=4.12.3
build>=1.2.1
cachetools>=5.4.0
certifi>=2024.7.4
cffi>=1.17.0
channels>=4.1.0
charset-normalizer>=3.3.2
chroma-hnswlib>=0.7.6
chromadb>=0.5.5
click>=8.1.7
coloredlogs>=15.0.1
colorlog>=6.8.2
comm>=0.2.2
constantly>=23.10.4
cryptography>=43.0.1
daphne>=4.1.2
dataclasses-json>=0.6.7
debugpy>=1.8.5
decorator>=5.1.1
Deprecated>=1.2.14
distro>=1.9.0
Django>=5.1
django-cors-headers>=4.4.0
django-storages>=1.14.4
executing>=2.0.1
fastapi>=0.112.0
filelock>=3.15.4
flatbuffers>=24.3.25
frozenlist>=1.4.1
fsspec>=2024.6.1
google-auth>=2.32.0
googleapis-common-protos>=1.63.2
greenlet>=3.0.3
grpcio>=1.65.4
h11>=0.14.0
httpcore>=1.0.5
httptools>=0.6.1
httpx>=0.27.0
huggingface-hub>=0.24.5
humanfriendly>=10.0
hyperlink>=21.0.0
idna>=3.7
importlib_metadata>=8.0.0
importlib_resources>=6.4.0
incremental>=24.7.2
ipykernel>=6.29.5
ipython>=8.26.0
jedi>=0.19.1
Jinja2>=3.1.3
jsonpatch>=1.33
jsonpointer>=3.0.0
jupyter_client>=8.6.2
jupyter_core>=5.7.2
kubernetes>=30.1.0
langchain>=0.2.12
langchain-community>=0.2.11
langchain-core>=0.2.28
langchain-openai>=0.1.20
langchain-text-splitters>=0.2.2
langsmith>=0.1.98
lxml>=5.2.2
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
marshmallow>=3.21.3
matplotlib-inline>=0.1.7
mdurl>=0.1.2
mmh3>=4.1.0
monotonic>=1.6
mpmath>=1.3.0
multidict>=6.0.5
mypy-extensions>=1.0.0
nest-asyncio>=1.6.0
networkx>=3.2.1
numpy>=1.26.4
nvidia-cublas-cu12>=12.4.2.65
nvidia-cuda-cupti-cu12>=12.4.99
nvidia-cuda-nvrtc-cu12>=12.4.99
nvidia-cuda-runtime-cu12>=12.4.99
nvidia-cudnn-cu12>=9.1.0.70
nvidia-cufft-cu12>=11.2.0.44
nvidia-curand-cu12>=10.3.5.119
nvidia-cusolver-cu12>=11.6.0.99
nvidia-cusparse-cu12>=12.3.0.142
nvidia-nccl-cu12>=2.20.5
nvidia-nvjitlink-cu12>=12.4.99
nvidia-nvtx-cu12>=12.4.99
oauthlib>=3.2.2
onnxruntime>=1.18.1
openai>=1.39.0
opentelemetry-api>=1.26.0
opentelemetry-exporter-otlp-proto-common>=1.26.0
opentelemetry-exporter-otlp-proto-grpc>=1.26.0
opentelemetry-instrumentation>=0.47b0
opentelemetry-instrumentation-asgi>=0.47b0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-proto>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-semantic-conventions>=0.47b0
opentelemetry-util-http>=0.47b0
orjson>=3.10.6
outcome>=1.3.0.post0
overrides>=7.7.0
packaging>=24.1
parso>=0.8.4
pexpect>=4.9.0
pillow>=10.2.0
platformdirs>=4.2.2
posthog>=3.5.0
prompt_toolkit>=3.0.47
protobuf>=4.25.4
psutil>=6.0.0
ptyprocess>=0.7.0
pure_eval>=0.2.3
pyasn1>=0.6.0
pyasn1_modules>=0.4.0
pycparser>=2.22
pydantic>=2.8.2
pydantic_core>=2.20.1
Pygments>=2.18.0
pyOpenSSL>=24.2.1
PyPika>=0.48.9
pyproject_hooks>=1.1.0
PySocks>=1.7.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
PyYAML>=6.0.1
pyzmq>=26.1.0
regex>=2024.7.24
requests>=2.32.3
requests-oauthlib>=2.0.0
rich>=13.7.1
rsa>=4.9
safetensors>=0.4.4
scipy>=1.14.0
selenium>=4.23.1
service-identity>=24.1.0
setuptools>=72.1.0
shellingham>=1.5.4
six>=1.16.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.5
SQLAlchemy>=2.0.32
sqlparse>=0.5.1
stack-data>=0.6.3
starlette>=0.37.2
sympy>=1.13.1
tenacity>=8.5.0
tiktoken>=0.7.0
tokenizers>=0.19.1
tornado>=6.4.1
tqdm>=4.66.5
traitlets>=5.14.3
transformers>=4.43.4
trio>=0.26.1
trio-websocket>=0.11.1
triton>=3.0.0
Twisted>=24.7.0
txaio>=23.1.1
typer>=0.12.3
typing-inspect>=0.9.0
typing_extensions>=4.12.2
urllib3>=2.2.2
uvicorn>=0.30.5
uvloop>=0.19.0
watchfiles>=0.22.0
wcwidth>=0.2.13
websocket-client>=1.8.0
websockets>=12.0
wheel>=0.43.0
wrapt>=1.16.0
wsproto>=1.2.0
yarl>=1.9.4
zipp>=3.19.2
zope.interface>=7.0.3
